refactor: barrel exporting implemented
