+++ app/api/user/snippets/[snippetId]/route.ts
@@ -1,7 +1,7 @@
 import { mongoConnect } from '@/utils';
 import { NextRequest, NextResponse } from 'next/server';
 import jwt from 'jsonwebtoken';
-import { Snippet } from '@/models/snippet';
+import { Snippet } from '@/models';
 
 export async function DELETE(
   req: NextRequest,
