+++ components/Provider.tsx
@@ -1,9 +1,8 @@
 'use client';
 
-import { AlertProvider } from '@/contexts/AlertContext';
-import { UserProvider } from '@/contexts/UserContext';
+import { AlertProvider, UserProvider } from '@/contexts';
 import { FC, ReactNode } from 'react';
-import { Alert } from './ui/Alert';
+import { Alert } from '@/components';
 
 interface IProvider {
   children: ReactNode;
