+++ components/index.ts
@@ -1,1 +1,6 @@
 export { Header } from './dashboard/Header';
+export { UserSnippets } from './UserSnippets';
+export { Editor } from './Editor';
+export { Snippet } from './Snippet';
+export { Alert } from './ui/Alert';
+export { SnippetCard } from './SnippetCard';
