+++ hooks/index.ts
@@ -1,2 +1,4 @@
 export { useTimeout } from './useTimeout';
 export { useUser } from './useUser';
+export { useError } from './useError';
+export { useAlert } from './useAlert';
