+++ hooks/useAlert.ts
@@ -1,4 +1,5 @@
-import { AlertContext, IAlertContext } from '@/contexts/AlertContext';
+import { AlertContext } from '@/contexts/AlertContext';
+import { IAlertContext } from '@/types';
 import { useContext } from 'react';
 
-export const useAlert = () => useContext(AlertContext) as IAlertContext
+export const useAlert = () => useContext(AlertContext) as IAlertContext;
