+++ models/index.ts
@@ -1,0 +1,2 @@
+export { Snippet } from './snippet';
+export { User } from './user';
