+++ schema/index.ts
@@ -1,0 +1,2 @@
+export type { ISnippet } from './snippet';
+export type { IUser } from './user';
