+++ types/IUserContext.d.ts
@@ -1,0 +1,4 @@
+export interface IUserContext {
+  user: IUser | null;
+  setUser: React.Dispatch<React.SetStateAction<IUser | null>>;
+}
