+++ types/index.ts
@@ -1,0 +1,5 @@
+export type { IUserContext } from './IUserContext';
+export type { IEditor } from './IEditor';
+export type { ISnippet } from './ISnippet';
+export type { IUser } from './IUser';
+export type { IAlertContext } from './IAlertContext';
