+++ utils/index.ts
@@ -1,3 +1,5 @@
 export { regex } from './regex';
 export { mongoConnect } from './mongoConnect';
 export { genPassword } from './genPassword';
+export { genToken } from './genToken';
+export { isPasswordValid } from './validator';
